pkg/13/main_test.go
package main

import (
	"strings"
	"testing"
)

func TestPart1(t *testing.T) {
	result, err := Part1(strings.Split(`#.##..##.
..#.##.#.
##......#
##......#
..#.##.#.
..##..##.
#.#.##.#.

#...##..#
#....#..#
..##..###
#####.##.
#####.##.
..##..###
#....#..#`, "\n"))
	if err != nil {
		t.Fatal(err)
	}
	if result != 405 {
		t.Fatalf("expected 405, got %d", result)
	}
}

func TestPart2(t *testing.T) {
	result, err := Part2(strings.Split(`#.##..##.
..#.##.#.
##......#
##......#
..#.##.#.
..##..##.
#.#.##.#.

#...##..#
#....#..#
..##..###
#####.##.
#####.##.
..##..###
#....#..#`, "\n"))
	if err != nil {
		t.Fatal(err)
	}
	if result != 400 {
		t.Fatalf("expected 400, got %d", result)
	}
}
